test: move truncation tests to their own file to avoid positional conflict with #2947
Both PRs appended sections at the same anchor in test_anthropic_model.py
and git misaligned the similar stub-class bodies into interleaving hunks.
With this PR's tests in test_anthropic_model_truncation.py the two
branches share no modified test file and merge clean in either order.
